Officially issue #558 on Github
This commit has changes to the Japan event txt file which will be elaborated on later on.

To make the new Japan events work correctly minor changes have been made in various other files, the decision file has been altered so that the 'buy back' decision only fires if Japan has cores and only has them on the southern Kuril islands. The province history files have also been modified so that Japan now starts with cores on only the 4 main islands, the territory they start with in the 1946 stardate, this to both represent them renouncing their authority over the islands owned by the US and to prevent the negatvie relation modifier by the US owning their cores.

Now about the main events txt file, the Japan independence event now is triggered by the San Francisco event instead of firing standalone, this because Japan officialy got unoccupied after the treaty and now that the treaty is added as an event it would only make sense to make it an followup event to it. The San Francisco treaty event also fires later on than the normal independence event could so Japan will have the puppet state prestige modifier for a longer period, slightly preventing them from slipping into GP status too early on. The transfer islands event have now been modified to add cores if the US transfers them back to Japan, if the US refuses to hand them back Japan gets an additional event to claim these islands once again. Also if Japan is a communist dictatorship these events will not fire for the US.

The new main even tree is a 'dynamic' way for a player to choose which islands north of Japan he would like to claim and which he would not, claiming nothing and thus not angering the Soviet wil help the player slightly if he wants too become a Commmunist Japan later on, choosing to either claim some or all islands will anger the soviets and lower their influence in Japan, representing that the US tried to keep soviet influence out of Japan post-WW2. If the player or the AI follows the historical outcome the decision to buy back the southern islands will work as normally, else the decision will not be available and the player will have to reaquire their claims through war.

I was thinking of making new events for a commie Japan to reclaim their southern islands if the US transfer events never fired but I was still unsure of how to do that.